Refactor all database locks.
The locks were not correctly wrapping the full transactions, nor were they
giving the right amount of consistency. It's less performant, but the most
reliable way to keep consistent state during transactions is to lock all rows
in a given 'table'. This commit makes that change, and also moves the locks
into the APIs so that they can correctly wrap the entire transactions.