Improved progress calculation. (close #54)
This change does a full history scan periodically of an archive to calculate
the exact current size and download state of the archive. It's not the *most*
efficient option; dat-node has a better iteration which employs diffing to
make progress checks faster. But, this should be reliable and easier to read.
We can optimize later.